Add new concept: LayerProvider Refactor LayerList and DescribeLayer to use available LayerProviders
